+++ lib/mc_markdown/renderers.rb
@@ -44,5 +44,3 @@ end
 
 require_relative 'renderers/html'
 require_relative 'renderers/base'
-require_relative 'renderers/guidebook'
-require_relative 'renderers/legal'