+++ lib/mc_markdown/renderers/guidebook.rb
@@ -1,5 +1,0 @@
-module MCMarkdown
-  class Guidebook < MCMarkdown::Base
-    include ::MCMarkdown::Formatter::HeaderWithID
-  end
-end